[Internal]Fixing  test regressions introduced by recent PR adding new Otel attributes (#45826)
* Fixing Otel attributes recently added

* Update CHANGELOG.md

* Update sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/TracerUnderTest.java


* Update sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/CosmosTracerTest.java


---------

+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/DiagnosticsProvider.java
@@ -1540,7 +1540,7 @@ public final class DiagnosticsProvider {
                 context);
             tracer.setAttribute(
                 "azure.cosmosdb.response.sub_status_code",
-                cosmosCtx.getSubStatusCode(),
+                (Integer)cosmosCtx.getSubStatusCode(),
                 context);
             tracer.setAttribute(
                 "db.cosmosdb.request_charge",
